+++ lib/models/x_axis_label_style.dart
@@ -2,16 +2,15 @@ import 'package:flutter/material.dart';
 
 class XAxisLabelStyle {
   final bool enabled;
-  final Color color;
-  final double fontSize;
-  final FontWeight? fontWeight;
+  final TextStyle textStyle;
   final double distanceFromAxis;
 
   const XAxisLabelStyle({
     this.enabled = true,
-    this.color = Colors.black87,
-    this.fontSize = 12.0,
-    this.fontWeight,
+    this.textStyle = const TextStyle(
+      color: Colors.black87,
+      fontSize: 12.0,
+    ),
     this.distanceFromAxis = 8.0,
   });
 }